Move all Ollama components into a new ollama directory for better organization. Add a new ModelsSelector component that we can re-use to avoid duplicate code
