+++ src/js/settings/components/provider-settings/ollama/multimodal.js
@@ -9,29 +9,37 @@ import { useEffect } from '@wordpress/element';
 /**
  * Internal dependencies
  */
-import { STORE_NAME } from '../../data/store';
-import { OllamaBaseSettings } from './ollama-base';
-import { SettingsRow } from '../settings-row';
+import { SettingsRow } from '../../settings-row';
+import { STORE_NAME } from '../../../data/store';
+import { OllamaBaseSettings } from './base';
+import { useFeatureContext } from '../../feature-settings/context';
+import { PromptRepeater } from '../../feature-additional-settings/prompt-repeater';
 
 /**
- * React Component for Ollama Embeddings settings.
+ * Component for Ollama Multimodal Provider settings.
  *
- * This component is used within the ProviderSettings component to
- * allow users to configure the Ollama Embeddings settings.
+ * This component is used within the ProviderSettings component
+ * to allow users to configure the Ollama Multimodal Provider settings.
  *
  * @param {Object}  props              Component props.
  * @param {boolean} props.isConfigured Whether the provider is configured.
  *
- * @return {React.ReactElement} OllamaEmbeddingsSettings component.
+ * @return {React.ReactElement} OpenAIChatGPTSettings component.
  */
-export const OllamaEmbeddingsSettings = ( { isConfigured = false } ) => {
-	const providerName = 'ollama_embeddings';
+export const OllamaMultimodalSettings = ( { isConfigured = false } ) => {
+	const { featureName } = useFeatureContext();
+	const providerName = 'ollama_multimodal';
 	const providerSettings = useSelect(
 		( select ) =>
 			select( STORE_NAME ).getFeatureSettings( providerName ) || {}
 	);
 	const { setProviderSettings } = useDispatch( STORE_NAME );
 	const onChange = ( data ) => setProviderSettings( providerName, data );
+	const setPrompts = ( prompts ) => {
+		setProviderSettings( providerName, {
+			prompt: prompts,
+		} );
+	};
 
 	const models = [
 		{ label: __( '-- Choose Model --', 'classifai' ), value: '' },
@@ -64,6 +72,24 @@ export const OllamaEmbeddingsSettings = ( { isConfigured = false } ) => {
 	// Check if we have actual models (not just the placeholder).
 	const hasModels = models.length > 1;
 
+	const promptExamples = (
+		<>
+			{ __( 'Add a custom prompt, if desired. See our', 'classifai' ) }
+			<a
+				href="https://10up.github.io/classifai/advanced-docs/prompt-examples"
+				target="_blank"
+				rel="noopener noreferrer"
+			>
+				{ ' ' }
+				{ __( 'documentation', 'classifai' ) }
+			</a>{ ' ' }
+			{ __(
+				'for some example prompts you can try that have been tested for specific use cases.',
+				'classifai'
+			) }
+		</>
+	);
+
 	return (
 		<>
 			<SettingsRow
@@ -88,6 +114,21 @@ export const OllamaEmbeddingsSettings = ( { isConfigured = false } ) => {
 					providerName={ providerName }
 					onChange={ onChange }
 				/>
+			) }
+			{ [
+				'feature_descriptive_text_generator',
+				'feature_image_to_text_generator',
+				'feature_image_tags_generator',
+			].includes( featureName ) && (
+				<SettingsRow
+					label={ __( 'Prompt', 'classifai' ) }
+					description={ promptExamples }
+				>
+					<PromptRepeater
+						prompts={ providerSettings.prompt }
+						setPrompts={ setPrompts }
+					/>
+				</SettingsRow>
 			) }
 		</>
 	);
