+++ src/js/settings/components/provider-settings/ollama/ollama.js
@@ -12,10 +12,10 @@ import { useEffect } from '@wordpress/element';
 /**
  * Internal dependencies
  */
-import { STORE_NAME } from '../../data/store';
-import { OllamaBaseSettings } from './ollama-base';
-import { useFeatureContext } from '../feature-settings/context';
-import { SettingsRow } from '../settings-row';
+import { STORE_NAME } from '../../../data/store';
+import { OllamaBaseSettings } from './base';
+import { useFeatureContext } from '../../feature-settings/context';
+import { SettingsRow } from '../../settings-row';
 
 /**
  * Component for Ollama Provider settings.
@@ -114,6 +114,7 @@ export const OllamaSettings = ( { isConfigured = false } ) => {
 						onChange={ ( value ) =>
 							onChange( { number_of_suggestions: value } )
 						}
+						__next40pxDefaultSize
 					/>
 				</SettingsRow>
 			) }
